requirements.txt
fastapi
uvicorn
streamlit
tensorflow
pillow
opencv-python
numpy
pandas